requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
Pillow==6.0.0
pip-autoremove==0.9.1
requests==2.21.0
text-editor==1.0.3
urllib3==1.24.2
